requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
backports.weakref==1.0.post1
boto3==1.28.50
botocore==1.31.50
brotlipy==0.7.0
constantly==15.1.0
cssselect==1.1.0
daal4py==2023.1.1
pandas==1.5.3
psycopg2-binary==2.9.7

